fix: amount type on RefuelingRequest

+++ src/main/java/br/thullyoo/desafio_tecnico_java/dto/RefuelingRequest.java
@@ -1,4 +1,6 @@
 package br.thullyoo.desafio_tecnico_java.dto;
 
-public record RefuelingRequest(Long fuelPumpId, Double amount, Long liters) {
+import java.math.BigDecimal;
+
+public record RefuelingRequest(Long fuelPumpId, BigDecimal amount, Long liters) {
 }
